Backend/requirements.txt
# ========== 🌐 Web 框架相关 ==========
fastapi==0.116.1               # 主框架
starlette==0.47.2              # FastAPI 的底层 ASGI 框架
uvicorn==0.35.0                # ASGI 服务器，用于运行 FastAPI

# ========== 🛡️ 安全认证 ==========
passlib[bcrypt]==1.7.4         # 用于密码哈希（bcrypt 支持）
python-jose==3.4.0             # JWT 编解码支持

# ========== 📦 数据库相关 ==========
SQLAlchemy==2.0.38             # ORM 框架
aiomysql==0.2.0                # 异步 MySQL 驱动

# ========== 📨 请求与文件处理 ==========
httpx==0.28.1                  # 异步 HTTP 客户端
python-multipart>=0.0.5        # 表单处理，FastAPI 上传文件需要
python-docx==1.2.0             # 处理 .docx 文件
PyMuPDF==1.23.9                # 处理 PDF 文件
Pillow==11.3.0                 # 图像处理库

# ========== 🧩 配置与环境变量 ==========
python-dotenv==1.1.1           # 读取 .env 配置文件
pydantic==2.11.7               # 数据验证模型
pydantic-settings==2.10.1      # Pydantic v2 的配置支持
email-validator==2.1.1         # Pydantic 验证 Email 时用到

# ========== 🧠 大模型生态 ==========
openai==1.97.1                 # OpenAI 官方 SDK
langchain==0.3.26              # LLM 应用开发框架
langchain-openai==0.3.28       # Langchain 的 OpenAI 集成
langchain-community            # Langchain 的社区工具集
dashscope==1.23.9              # 通义千问 API SDK
volcengine-python-sdk==4.0.6   # 火山引擎 SDK，用于调用其大模型或服务

# ========== 🔍 搜索工具 ==========
google-search-results          # SerpAPI 所需库，用于 Google 搜索
langchain_community            # 提供 SerpAPIWrapper 等工具（已列，无需重复）

# ========== 🕒 时间处理 ==========
pytz==2025.2                   # 时区支持（推荐替代为 `zoneinfo` + `pendulum`）
requests==2.32.4               # 同步 HTTP 请求（如果仅用 httpx 可考虑移除）
